[REFACTOR]: user 도메인 dto user, customs로 분리 후 이동
